libnm/match: support backslash escaping the modifier in match patterns
Without it, it's not possible to define a positive match that starts
with an exclamation mark. While of course it may not be very useful,
I think our configuration mechanism should allow to express every
possible value. Do this, by special casing a '\\' at the beginning
of the pattern (or right after the '!').

That means, "\\!a" now indicates the positive pattern "!a".